Revert "add Tests safety"
